+++ src/main/java/group/ship/blackshipstore/controller/OrderController.java
@@ -4,9 +4,15 @@ import group.ship.blackshipstore.entity.Order;
 import group.ship.blackshipstore.sevices.OrderService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 @RestController
 @RequestMapping("/order")
 @Tag(name = "Order Controller", description = "Controller to get Pirate Orders")
@@ -31,7 +37,7 @@ public class OrderController {
 
     @PutMapping("/{id}")
     @Operation(summary = "Mark Order completed", description = "Return Order that has been marked as completed by id")
-    public Order markOrderAsCompleted(@PathVariable Integer id) {
+    public Order markOrderAsCompleted(@PathVariable Long id) {
         return orderService.markOrderAsCompleted(id);
     }
