+++ src/main/java/group/ship/blackshipstore/repositories/ArticlesRepository.java
@@ -1,10 +1,9 @@
 package group.ship.blackshipstore.repositories;
 
 import group.ship.blackshipstore.entity.Article;
-import group.ship.blackshipstore.entity.ItemValues;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticlesRepository extends JpaRepository<Article,Integer> {
+public interface ArticlesRepository extends JpaRepository<Article, Long> {
 }
