+++ src/main/java/group/ship/blackshipstore/repositories/Repository.java
@@ -4,5 +4,5 @@ import group.ship.blackshipstore.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @org.springframework.stereotype.Repository
-public interface Repository extends JpaRepository<Item,Integer> {
+public interface Repository extends JpaRepository<Item, Long> {
 }
